feat: add Flight Stop

+++ backend/src/main/java/com/flight/backend/entity/FlightStop.java
@@ -43,7 +43,7 @@ public class FlightStop {
     private LocalDateTime departureTime;
 
     @Column(name = "stop_duration")
-    private Integer stopDuration;
+    private Long stopDuration;
 
     @Column(name = "stop_order")
     private int stopOrder;
